pyproject.toml
[project]
name = "agro-esg-carbon-compliance"
version = "0.1.0"
description = "AgroESG Project - Carbon Compliance & Risk Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- Core Engine (O "Motor" do ELT) ---
    "apache-airflow[google,postgres]==2.10.0",
    "psycopg2-binary>=2.9.0",
    "duckdb>=1.0.0",
    "pyarrow>=15.0.0",
    # --- Cloud & Storage (A "Infra") ---
    "google-cloud-storage",
    "google-cloud-bigquery",
    # --- Transformation (A "Inteligência") ---
    "dbt-bigquery>=1.7.0",
    "astronomer-cosmos[google]>=1.2.0",
    # --- Utils Essenciais ---
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "snowplow-tracker<1.0.0",
    "streamlit>=1.54.0",
    "plotly>=6.5.2",
    "folium>=0.20.0",
    "streamlit-folium>=0.26.1",
    "db-dtypes>=1.5.0",
    "pydeck>=0.9.1",
    "earthengine-api>=0.1.300",
    "geemap>=0.30.0",
    "fastapi>=0.131.0",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
]

[tool.uv]
managed = true

# Grupo de desenvolvimento
[dependency-groups]
dev = [
    "jupyterlab>=4.0.0",
    "black",
    "ruff",
]
